Remove old data validation button and add separate validation buttons
